Implement new feature to convert JSON to csv or parquet
This feature give you the ability to convert the JSON Hasty annotation file into a csv or parquet file. This can facilitate data retrieval for your pipeline.

The Hasty JSON file is still uploaded to S3 for data backup and ensuring your original annotation file is still available to you as it contains more information pertaining to Hasty.